moving tests to the same file. still some config mixup issue
